Add impersonation guards into auth utils
